support standalone boot

+++ tools/HiveConnectTest/hserver.js
@@ -1,5 +1,5 @@
 'use strict';
-/*
+
 var	HRENDER = '../../build/bin/hrender',
 	HRENDER_ARG = ['../../build/bin/HIVE_UI/hrender_server_ipc.lua', '--opengl'],
 	spawn = require('child_process').spawn;
@@ -7,6 +7,7 @@ var	HRENDER = '../../build/bin/hrender',
 
 function startupHRenderServer() {
 	'use strict';
+	var exitflag = false;
 	try {
 		var process = spawn(HRENDER, HRENDER_ARG);
 		process.stdout.on('data', function (data) {
@@ -18,7 +19,9 @@ function startupHRenderServer() {
 		process.on('exit', function (code) {
 			console.error('-------------------------\nhrender is terminated.\n-------------------------');
 			console.log('exit code: ' + code);
-			startupHRenderServer(); // reboot
+			if (!exitflag) {
+				startupHRenderServer(); // reboot
+			}
 		});
 		process.on('error', function (err) {
 			console.log('process error', err);
@@ -26,9 +29,15 @@ function startupHRenderServer() {
 	} catch (e) {
 		console.log('process error', e);
 	}
+	
+	function stopHRenderServer() {
+		exitflag = true;
+		process.kill();
+	}
+	module.exports.stopHRenderServer = stopHRenderServer;
 }
 startupHRenderServer();
-*/
+
 
 
 //-------------
